Реализован метод squareSequenceDigit() из lesson3.task1, доделан lesson1

+++ src/lesson1/task1/Simple.kt
@@ -80,7 +80,7 @@ fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double =
  * Вывести значение того же угла в радианах (например, 0.63256).
  */
 fun angleInRadian(deg: Int, min: Int, sec: Int): Double =
-    (deg + min / 60 + sec / 3600) * (PI / 180)
+    (deg.toDouble() + min.toDouble() / 60 + sec.toDouble() / 3600) * (PI / 180)
 
 /**
  * Тривиальная
